Add apiKey authentication structure
