+++ lib/Security.js
@@ -1,47 +1,11 @@
 'use strict'
 
-const templates = require('./templates')
+const jwt = require('./security/jwt')
+const apiKey = require('./security/apiKey')
 
 const Security = (service, commands) => {
-  const cleanUserData = userData => ({
-    _id: userData._id,
-    name: userData.name,
-    email: userData.email,
-    role: userData.role
-  })
-
-  const getUserByRefreshToken = refreshToken => commands.refreshToken.getUser(refreshToken)
-    .then(userData => Promise.resolve({
-      userData: cleanUserData(userData)
-    }))
-
-  const getUserByCredentials = userCredentials => {
-    return commands.user.get({
-      email: userCredentials.user,
-      password: userCredentials.password
-    }).then(userData => commands.refreshToken.add(userData)
-      .then(refreshToken => Promise.resolve({
-        userData: cleanUserData(userData),
-        refreshToken: refreshToken.token
-      })))
-  }
-
-  const authenticateHandler = body => {
-    const action = body.refreshToken ? getUserByRefreshToken(body.refreshToken) : getUserByCredentials(body)
-    return action.catch(error => service.tracer.error(templates.authenticationError({
-      message: error.message
-    })).then(() => Promise.reject(error)))
-  }
-
-  const revokeAuth = (userData, params, body) => {
-    if (userData.role === 'admin') {
-      return true
-    }
-    return getUserByRefreshToken(body.refreshToken)
-      .then(tokenUserData => tokenUserData.userData.email === userData.email ? Promise.resolve() : Promise.reject(new Error()))
-  }
-
-  const revokeHandler = body => commands.refreshToken.remove(body.refreshToken)
+  const jwtMethods = jwt.Methods(service, commands)
+  const apiKeyMethods = apiKey.Methods(service, commands)
 
   return {
     methods: {
@@ -49,11 +13,22 @@ const Security = (service, commands) => {
         secret: 'thisIsNotaRealTokenSecretPleaseReplaceIt',
         expiresIn: 180,
         authenticate: {
-          handler: authenticateHandler
+          handler: jwtMethods.authenticateHandler
         },
         revoke: {
-          auth: revokeAuth,
-          handler: revokeHandler
+          auth: jwtMethods.revokeAuth,
+          handler: jwtMethods.revokeHandler
+        }
+      },
+      apiKey: {
+        verify: apiKeyMethods.verify,
+        authenticate: {
+          auth: apiKeyMethods.authenticateAuth,
+          handler: apiKeyMethods.authenticateHandler
+        },
+        revoke: {
+          auth: apiKeyMethods.revokeAuth,
+          handler: apiKeyMethods.revokeHandler
         }
       }
     }
