+++ lib/security/apiKey.js
@@ -1,0 +1,27 @@
+'use strict'
+
+// const templates = require('../templates')
+
+const Methods = (service, commands) => {
+  const verify = () => {}
+
+  const authenticateAuth = () => {}
+
+  const authenticateHandler = () => {}
+
+  const revokeAuth = () => {}
+
+  const revokeHandler = () => {}
+
+  return {
+    verify,
+    authenticateAuth,
+    authenticateHandler,
+    revokeAuth,
+    revokeHandler
+  }
+}
+
+module.exports = {
+  Methods
+}
